+++ src/backend/marsha/bbb/apps.py
@@ -10,3 +10,9 @@ class BbbConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "marsha.bbb"
     verbose_name = _("Big Blue Button")
+
+    def ready(self):
+        # Signals must be imported and connected once the app is ready.
+        # Callbacks are connected thanks to the "receiver" decorator.
+        # pylint: disable=import-outside-toplevel, unused-import
+        import marsha.bbb.signals  # noqa